+++ src/nav-list/nav-list.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'ign-nav-list',
+    template: `<mat-nav-list><ng-content></ng-content></mat-nav-list>`,
+})
+export class NavListComponent {}